Minor code fixes: final variables, javadoc fixes, Iterator -> for : loops, type casting, primitive unboxing, unnecessary 'else'.

+++ metafacades/uml/emf/uml22/src/main/java/org/andromda/metafacades/emf/uml22/EntityAssociationEndLogicImpl.java
@@ -74,7 +74,7 @@ public class EntityAssociationEndLogicImpl
                     columnNamePrefix,
                     this,
                     UMLProfile.TAGGEDVALUE_PERSISTENCE_COLUMN,
-                    ((Entity)this.getType()).getMaxSqlNameLength(),
+                    Short.valueOf(((Entity)this.getType()).getMaxSqlNameLength()),
                     this.getForeignKeySuffix(),
                     this.getConfiguredProperty(UMLMetafacadeProperties.SQL_NAME_SEPARATOR));
         }
